+++ ft_memset.c
@@ -6,13 +6,14 @@
 /*   By: idavoli- <idavoli-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/22 15:20:26 by idavoli-          #+#    #+#             */
-/*   Updated: 2021/08/24 23:33:46 by idavoli-         ###   ########.fr       */
+/*   Updated: 2021/08/30 14:38:38 by idavoli-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stddef.h>
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, unsigned int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
 	unsigned char	*p;
